move sot related code in a separate dir
