+++ paddle/fluid/pybind/pybind.cc
@@ -1098,9 +1098,7 @@ PYBIND11_MODULE(libpaddle, m) {
   BindCudaStream(&m);
   BindXpuStream(&m);
   BindJit(&m);
-#if EVAL_FRAME_IS_SUPPORTED
-  BindEvalFrame(&m);
-#endif
+  BindSOT(&m);
   BindCustomDevicePy(&m);
   BindEagerUtils(m.ptr());
